requirements.txt
requests~=2.31.0
requests_cache
icecream~=2.1.3
numpy~=1.26.2
matplotlib
pytest~=7.4.3
python-dotenv~=1.0.0
flask~=3.0.0